Transaction is now exposed as TransactionContext instead of ByteBuffer.
