+++ azure-servicebus/src/main/java/com/microsoft/azure/servicebus/IMessageAndSessionPump.java
@@ -8,7 +8,6 @@ import java.util.UUID;
 import java.util.concurrent.CompletableFuture;
 
 import com.microsoft.azure.servicebus.primitives.ServiceBusException;
-import com.microsoft.azure.servicebus.security.TransactionContext;
 
 /**
  * Represents the pump which is underneath the clients that handles message processing.
