+++ azure-servicebus/src/main/java/com/microsoft/azure/servicebus/QueueClient.java
@@ -11,7 +11,6 @@ import java.util.Map;
 import java.util.UUID;
 import java.util.concurrent.CompletableFuture;
 
-import com.microsoft.azure.servicebus.security.TransactionContext;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
