+++ azure-servicebus/src/main/java/com/microsoft/azure/servicebus/TopicClient.java
@@ -8,7 +8,6 @@ import java.time.Instant;
 import java.util.Collection;
 import java.util.concurrent.CompletableFuture;
 
-import com.microsoft.azure.servicebus.security.TransactionContext;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
