+++ azure-servicebus/src/test/java/com/microsoft/azure/servicebus/QueueSessionTests.java
@@ -2,11 +2,9 @@ package com.microsoft.azure.servicebus;
 
 import com.microsoft.azure.servicebus.primitives.MessageNotFoundException;
 import com.microsoft.azure.servicebus.primitives.ServiceBusException;
-import com.microsoft.azure.servicebus.security.TransactionContext;
 import org.junit.Assert;
 import org.junit.Test;
 
-import java.nio.ByteBuffer;
 import java.util.UUID;
 import java.util.concurrent.ExecutionException;
